Create applicant_scripts_Section1_dd.py
Python code will take a directory of reports and summarize the average fraction of total reads per species. The first argument is the directory, the second is a threshold (default is 0.5%) and the third is the output file name in tab-separated format. 

The arguments in this script were run on a local machine and should be changed for use on another machine.