Create applicant_scripts_Section2_dd.sh
The following Bash script will take input from a Staphylococcus aureus assembly genome to BLAST against a query fasta and output a file with 100% query coverage. Since the assembly given was a DNA sequence rather than a protein sequence, the script translates the assembly to a protein sequence using bacterial codon rules and assuming the frame starts at the first nucleotide. 

The example given assumes that the DNA assembly is in "assembly/" the query sequences are in "query/" and the tab-separated output file will be "hits.txt" and was run on a local machine. Chance directories and output file name as needed.